Chatbot/requirements.txt
langgraph
langchain
langchain-groq
langchain-community
tavily-python
beautifulsoup4
python-dotenv
